honestspace_backend/requirements.txt
# Core Django
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=23.0
psycopg2-binary>=2.9.0
django-extensions>=3.2.0

# JWT Authentication
djangorestframework-simplejwt>=5.2.0
PyJWT>=2.6.0

# Social Authentication (Google OAuth)
social-auth-app-django>=5.2.0

# Database and Caching
redis>=4.5.0
django-redis>=5.2.0

# PostgreSQL with PostGIS support
psycopg2>=2.9.0

# API Documentation
drf-yasg>=1.21.0

# File handling and storage
Pillow>=9.5.0
django-storages>=1.13.0
boto3>=1.26.0  # For AWS S3

# Background tasks
celery>=5.2.0
django-celery-beat>=2.5.0

# Environment variables
python-decouple>=3.8
python-dotenv>=1.0.0

# Payment gateways
stripe>=5.4.0
paypalrestsdk>=1.13.0

# Utilities
requests>=2.28.0
phonenumbers>=8.13.0
python-slugify>=8.0.0

# Geographic features
geodjango

# Testing
pytest>=7.3.0
pytest-django>=4.5.0
factory-boy>=3.2.0

# Development tools
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Monitoring and logging
sentry-sdk>=1.21.0

# Email
sendgrid>=6.10.0

# Image processing
Pillow>=9.5.0
pillow-heif>=0.10.0

# Data validation
marshmallow>=3.19.0
django-phonenumber-field>=7.1.0

# API rate limiting
django-ratelimit>=4.0.0

# Timezone handling
pytz>=2023.3

# JSON handling
ujson>=5.7.0

# HTTP client
httpx>=0.24.0